# apps/server/test/unit/common/__snapshots__/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tenants": {
      "get": {
        "operationId": "TenantController_listTenants",
        "summary": "获取租户列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "TenantController_listUsers",
        "summary": "获取租户内用户列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "TenantController_createUser",
        "summary": "创建用户",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{userId}": {
      "put": {
        "operationId": "TenantController_updateUser",
        "summary": "更新用户",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/roles": {
      "get": {
        "operationId": "TenantController_listRoles",
        "summary": "获取租户内角色列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "TenantController_createRole",
        "summary": "创建角色",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/roles/{roleId}": {
      "put": {
        "operationId": "TenantController_updateRole",
        "summary": "更新角色",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "TenantController_deleteRole",
        "summary": "删除角色",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/my-authorizers": {
      "get": {
        "operationId": "TenantController_getMyAuthorizers",
        "summary": "获取当前用户可管理的公众号",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/my-subscription": {
      "get": {
        "operationId": "TenantController_getMySubscription",
        "summary": "当前租户的订阅信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/plans": {
      "get": {
        "operationId": "TenantController_listPlans",
        "summary": "获取订阅套餐列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/permissions": {
      "get": {
        "operationId": "TenantController_listPermissions",
        "summary": "获取所有可用权限列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "团队管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "WXGZH",
    "description": "",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}